# src/main/java/com/demo/dto/ResponseDepartmentBean.java
package com.demo.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ResponseDepartmentBean {
    private Long id;
    private String name;
    private Long employeeCount;
}